maint: move state function to state.py
